Add storage API for common actions
This API stores data in an IndexedDB database called 'Questions'. This database consists of two object stores: 'questions' and 'comments'.

IndexedDB (with 'idb' wrapper) was chosen because it's available out of the box on majority of supported browsers and, unlike LocalStorage, has a good support for storing JavaScript object data.